pyproject.toml
[project]
name = "llama-pan-latam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.2",
    "librosa>=0.10.2.post1",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.1",
    "soundfile>=0.12.1",
    "streamlit>=1.40.1",
    "wavio>=0.0.9",
]
